Add files via upload
I have created a table in PostgreSQL called transactions, and I will be importing the data from our CSV file "Jovis Online Retail" into this table. It is usually at this point that I encounter problems with some CSV files because of their data format. Postgres is very strict about formatting and will throw an error if it notices any issues with the file or the format.